requirements.txt
paddlepaddle-gpu==2.6.1
tqdm==4.66.4
numpy==1.23.5
torch==2.3.1+cu121
rich==13.7.1
pyyaml